Remove text-size-adjust
That breaks the font-size zoom setting on Safari on iOS now, and won’t work any longer on iPadOS when the desktop-class interference kicks in. Actually, it seems completely broken in that case, as it disables the interference for some dom elements in some books but not others, etc.

This means we must find a way to handle this case, as well as Firefox issues.